# backend/src/docs/api-docs.yaml
openapi: 3.1.0
info:
  title: "Express server with typescript"
  version: 0.1.1
  summary: "Express server"
  description: "Training API for working with products. Data is stored in JSON (no database)."
servers:
  - url: http://localhost:3000

paths:
  /:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: The server is alive
          content:
            text/plain:
              schema:
                type: string
                example: "Express API is running. Try /api/products"

  /api/products:
    get:
      summary: Get a list of products
      responses:
        "200":
          description: Array of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value:
                    error: "Internal Server Error"

    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Incorrect input (e.g., title is empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductErrorResponse"
              examples:
                titleRequired:
                  value:
                    error: "Validation failed"
                    details:
                      - field: "title"
                        message: "Title is required and must be a string"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value:
                    error: "Internal Server Error"

  /api/products/{id}:
    get:
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the product (string)
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: "Not found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value:
                    error: "Internal Server Error"

    patch:
      summary: Partial product update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the product (string)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: "Not found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value:
                    error: "Internal Server Error"

    delete:
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the product (string)
      responses:
        "200":
          description: Successfully deleted product
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
              example:
                ok: true
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: "Not found"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal:
                  value:
                    error: "Internal Server Error"

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          example: "ab12cd34"
        title:
          type: string
          example: "Headphones"
        category:
          type: string
          example: "Gadgets"
        description:
          type: string
          example: "A very good Headphones"
        price:
          type: number
          exclusiveMinimum: 0
          example: 14.69
        stock:
          type: number
          example: 12
        rating:
          type: number
          example: 4.2
        imageUrl:
          type: string
          example: "https://www.example.com/image.jpg"
      required:
        - id
        - title
        - category
        - description
        - price
        - stock
        - rating

    ProductCreate:
      type: object
      properties:
        title:
          type: string
          example: "Headphones"
        category:
          type: string
          example: "Gadgets"
        description:
          type: string
          example: "A very good Headphones"
        price:
          type: number
          exclusiveMinimum: 0
          example: 14.69
        stock:
          type: number
          example: 12
        rating:
          type: number
          example: 4.2
        imageUrl:
          type: string
          example: "https://www.example.com/image.jpg"
      required:
        - title
        - category
        - description
        - price
        - stock
        - rating

    ProductPatch:
      type: object
      description: Partial update — all fields are optional
      properties:
        title:
          type: string
          example: "Headphones"
        category:
          type: string
          example: "Gadgets"
        description:
          type: string
          example: "A very good Headphones"
        price:
          type: number
          exclusiveMinimum: 0
          example: 14.69
        stock:
          type: number
          example: 12
        rating:
          type: number
          example: 4.2
        imageUrl:
          type: string
          example: "https://www.example.com/image.jpg"

    ProductErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "title"
              message:
                type: string
                example: "Title is required and must be a string"
      required:
        - error
        - details

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Internal Server Error"
      required:
        - error